requirements.txt
aioboto3==13.1.1
aiobotocore==2.13.1
aiofiles==24.1.0
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
attrs==23.2.0
betterlogging==1.0.0
boto3==1.34.131
botocore==1.34.131
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iso8601==1.1.0
Jinja2==3.1.4
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
pypika-tortoise==0.1.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
s3transfer==0.10.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tortoise-orm==0.21.5
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.3
watchfiles==0.22.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
